WitcherNx
Hey data nerds 👋, in this video we made an attempt to extract relationships between The Witcher's characters from the books. We also did some graph analyses (centrality measures and community detection) and visualization using NetworkX and Pyvis. The analyses are still quite preliminary and I might have glossed over some details, but hopefully, this is a bit interesting and you learned something from it 🤓. I hope you enjoyed this Viz and any questions and thoughts are welcome! Thank you for watching 🙌🏽